[MV] added support for giving out apiVersions supported response
taking the api version from the input and if the apiVersion is > 4 we give out errorCode 35. If the apiVersions is 0-4, we give it back as response. first 4 bytes - message length. next 4 bytes corrID. next 4 bytes - errorCode. next we are replying with apiKeysCount. here we are replying with apiVersions supported but we can also add other features if needed. apiVersion reply is 18 (2 bytes) and then next 2 byte 0 - minVersion supported, next 2 bytes -> 4 - maxVersion supported